feat: add example section with translation
