--- app/page.tsx
@@ -1,9 +1,0 @@
-import { ExampleSection } from '@/sections/homeSections/ExampleSection';
-
-export default function Home() {
-  return (
-    <main>
-      <ExampleSection />
-    </main>
-  );
-}
